# PI/CogniKids/cognikids-backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CogniKids API",
    "description": "Documentação interativa da API CogniKids.",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": ["http"],
  "paths": {
...
    "/classes": {
      "post": {
        "summary": "Criar nova turma",
        "parameters": [
          {"name": "body", "in": "body", "required": true, "schema": {"type": "object"}}
        ],
        "responses": {"201": {"description": "Turma criada"}}
      },
      "get": {
        "summary": "Listar turmas",
        "responses": {"200": {"description": "Lista de turmas"}}
      }
    },
    "/classes/{class_id}": {
      "get": {
        "summary": "Obter detalhes de uma turma",
        "parameters": [
          {"name": "class_id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {"200": {"description": "Detalhes da turma"}}
      }
    },
...
  }
}